Silence various compiler warnings

+++ avs_core/convert/convert_bits.cpp
@@ -1106,7 +1106,7 @@ AVSValue __cdecl ConvertBits::Create(AVSValue args, void* user_data, IScriptEnvi
     auto frame0 = clip->GetFrame(0, env);
     const AVSMap* props = env->getFramePropsRO(frame0);
     if (env->propNumElements(props, "_ColorRange") > 0) {
-      ColorRange_src = env->propGetInt(props, "_ColorRange", 0, nullptr); // fixme: range check
+      ColorRange_src = (int)env->propGetInt(props, "_ColorRange", 0, nullptr); // fixme: range check
     }
     else {
       // no param, no frame property -> rgb is full others are limited
